Extract daemon statsCollector to its own package
